Generic CommandBuffer<T>, and processing_utils crate
