+++ Cargo.toml
@@ -25,6 +25,7 @@ bevy = { git = "https://github.com/bevyengine/bevy", branch = "main" }
 processing = { path = "." }
 processing_pyo3 = { path = "crates/processing_pyo3" }
 processing_render = { path = "crates/processing_render" }
+processing_utils = { path = "crates/processing_utils" }
 
 [dependencies]
 bevy = { workspace = true }
