+++ crates/processing_render/src/graphics.rs
@@ -29,12 +29,11 @@ use crate::{
     Flush,
     error::{ProcessingError, Result},
     image::{Image, bytes_to_pixels, create_readback_buffer, pixel_size, pixels_to_bytes},
-    render::{
-        RenderState,
-        command::{CommandBuffer, DrawCommand},
-    },
+    render::{RenderState, command::DrawCommand},
     surface::Surface,
 };
+
+use processing_utils::CommandBuffer;
 
 pub struct GraphicsPlugin;
 
@@ -242,7 +241,7 @@ pub fn create(
             }),
             Transform::from_xyz(0.0, 0.0, 999.9),
             render_layer,
-            CommandBuffer::new(),
+            CommandBuffer::<DrawCommand>::new(),
             RenderState::default(),
             SurfaceSize(width, height),
             Graphics {
@@ -465,7 +464,7 @@ pub fn end_draw(app: &mut App, entity: Entity) -> Result<()> {
 
 pub fn record_command(
     In((graphics_entity, cmd)): In<(Entity, DrawCommand)>,
-    mut graphics_query: Query<&mut CommandBuffer>,
+    mut graphics_query: Query<&mut CommandBuffer<DrawCommand>>,
 ) -> Result<()> {
     let mut command_buffer = graphics_query
         .get_mut(graphics_entity)
