+++ crates/processing_render/src/render/command.rs
@@ -38,24 +38,3 @@ pub enum DrawCommand {
     },
     Geometry(Entity),
 }
-
-#[derive(Debug, Default, Component)]
-pub struct CommandBuffer {
-    pub commands: Vec<DrawCommand>,
-}
-
-impl CommandBuffer {
-    pub fn new() -> Self {
-        Self {
-            commands: Vec::new(),
-        }
-    }
-
-    pub fn push(&mut self, cmd: DrawCommand) {
-        self.commands.push(cmd);
-    }
-
-    pub fn clear(&mut self) {
-        self.commands.clear();
-    }
-}
