+++ crates/processing_utils/Cargo.toml
@@ -1,0 +1,10 @@
+[package]
+name = "processing_utils"
+version = "0.1.0"
+edition = "2024"
+
+[dependencies]
+bevy = { workspace = true }
+
+[lints]
+workspace = true
